Move Extensions to VideoExtensions
