+++ VideoExport.Core/VideoExtensions/MKVExtension.cs
@@ -2,7 +2,7 @@
 using UnityEngine;
 using VideoExport.ScreenshotPlugins;
 
-namespace VideoExport.Extensions
+namespace VideoExport.VideoExtensions
 {
     public class MKVExtension : AFFMPEGBasedExtension
     {
