+++ VideoExport.Core/VideoExtensions/MOVExtension.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 using VideoExport.ScreenshotPlugins;
 
-namespace VideoExport.Extensions
+namespace VideoExport.VideoExtensions
 {
     public class MOVExtension : AFFMPEGBasedExtension
     {
