+++ VideoExport.Core/VideoExtensions/MP4Extension.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using UnityEngine;
 using VideoExport.ScreenshotPlugins;
 
-namespace VideoExport.Extensions
+namespace VideoExport.VideoExtensions
 {
     public class MP4Extension : AFFMPEGBasedExtension
     {
